+++ app/client/src/PluginActionEditor/components/PluginActionResponse/utils/actionHasFailed.ts
@@ -1,0 +1,9 @@
+import type { ActionResponse } from "api/ActionAPI";
+
+function hasFailed(actionResponse: ActionResponse) {
+  return actionResponse.statusCode
+    ? actionResponse.statusCode[0] !== "2"
+    : false;
+}
+
+export default hasFailed;
