+++ app/client/src/components/editorComponents/JSResponseView.tsx
@@ -29,11 +29,11 @@ import EntityBottomTabs from "./EntityBottomTabs";
 import { getIsSavingEntity } from "selectors/editorSelectors";
 import { getJSResponseViewState } from "./utils";
 import { getFilteredErrors } from "selectors/debuggerSelectors";
+import { NoResponse } from "PluginActionEditor/components/PluginActionResponse/components/NoResponse";
 import {
-  NoResponse,
   ResponseTabErrorContainer,
   ResponseTabErrorContent,
-} from "./ApiResponseView";
+} from "PluginActionEditor/components/PluginActionResponse/components/ApiResponse";
 import LogHelper from "./Debugger/ErrorLogs/components/LogHelper";
 import LOG_TYPE from "entities/AppsmithConsole/logtype";
 import type { Log, SourceEntity } from "entities/AppsmithConsole";
@@ -45,7 +45,7 @@ import { EditorViewMode } from "ee/entities/IDE/constants";
 import ErrorLogs from "./Debugger/Errors";
 import { isBrowserExecutionAllowed } from "ee/utils/actionExecutionUtils";
 import JSRemoteExecutionView from "ee/components/JSRemoteExecutionView";
-import { IDEBottomView, ViewHideBehaviour } from "../../IDE";
+import { IDEBottomView, ViewHideBehaviour } from "IDE";
 
 const ResponseTabWrapper = styled.div`
   display: flex;
@@ -229,8 +229,8 @@ function JSResponseView(props: Props) {
                   <>
                     {responseStatus === JSResponseState.NoResponse && (
                       <NoResponse
-                        isButtonDisabled={disabled}
-                        isQueryRunning={isLoading}
+                        isRunDisabled={disabled}
+                        isRunning={isLoading}
                         // eslint-disable-next-line @typescript-eslint/ban-ts-comment
                         // @ts-ignore
                         onRunClick={onButtonClick}
